+++ src/components/inference/service-card.module.css
@@ -111,6 +111,25 @@
   font-size: 12px;
 }
 
+/* Templates only: what the bundle pre-downloads. Boxed off so a card that has it still reads as the
+   same card, with one extra block, rather than as a different kind of tile. */
+
+.included {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+  padding: 10px;
+  border-radius: 12px;
+  background: var(--background-glass-secondary);
+}
+
+.includedHead {
+  color: var(--text-secondary);
+  font-size: 11px;
+  letter-spacing: 0.04em;
+  text-transform: uppercase;
+}
+
 .metaRow {
   display: flex;
   flex-wrap: wrap;
