--- src/components/inference/template-card.module.css
@@ -1,139 +1,0 @@
-/* Sibling of template-card.module.css — same tile/rhythm so services and bundles read as one system.
-   `--accent` is set inline per card from the category meta. */
-.card {
-  text-align: left;
-  transition: transform 0.3s ease;
-
-  &:hover {
-    transform: scale(1.01);
-  }
-
-  &:focus-visible {
-    outline: 2px solid var(--accent1);
-    outline-offset: 2px;
-  }
-}
-
-.cardTop {
-  display: flex;
-  align-items: center;
-  gap: 12px;
-}
-
-.tile {
-  display: grid;
-  place-items: center;
-  flex: 0 0 38px;
-  width: 38px;
-  height: 38px;
-  border-radius: 12px;
-  color: var(--accent);
-  background: color-mix(in srgb, var(--accent) 12%, transparent);
-  box-shadow: inset 0 0 0 1px color-mix(in srgb, var(--accent) 30%, transparent);
-}
-
-.tileIcon,
-.tileLogo {
-  grid-area: 1 / 1;
-}
-
-.tileIcon {
-  width: 20px;
-  height: 20px;
-  font-size: 20px;
-  fill: currentColor;
-}
-
-.tileLogo {
-  width: 22px;
-  height: 22px;
-  object-fit: contain;
-}
-
-.tileMono {
-  font-family: var(--font-plus-jakarta-sans), var(--font-inter), sans-serif;
-  font-size: 12.5px;
-  font-weight: 700;
-}
-
-.titleWrap {
-  display: flex;
-  flex: 1;
-  flex-direction: column;
-  gap: 1px;
-  min-width: 0;
-}
-
-/* The outcome is the headline — two lines, then clamp. */
-.outcome {
-  display: -webkit-box;
-  overflow: hidden;
-  font-size: 14.5px;
-  font-weight: 700;
-  line-height: 1.25;
-  overflow-wrap: anywhere;
-  -webkit-box-orient: vertical;
-  -webkit-line-clamp: 2;
-}
-
-.service {
-  overflow: hidden;
-  color: var(--text-secondary);
-  font-size: 12.5px;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-}
-
-.included {
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-  padding: 10px;
-  border-radius: 12px;
-  background: var(--background-glass-secondary);
-}
-
-.includedHead {
-  color: var(--text-secondary);
-  font-size: 11px;
-  letter-spacing: 0.04em;
-  text-transform: uppercase;
-}
-
-.metaRow {
-  display: flex;
-  flex-wrap: wrap;
-  align-items: center;
-  gap: 12px;
-  color: var(--text-secondary);
-  font-size: 11.5px;
-  font-variant-numeric: tabular-nums;
-}
-
-.chip {
-  display: inline-flex;
-  align-items: center;
-  gap: 5px;
-  font-size: 12px;
-}
-
-.chipIcon {
-  width: 14px;
-  height: 14px;
-  font-size: 14px;
-  fill: currentColor;
-}
-
-.meta {
-  display: inline-flex;
-  align-items: center;
-  gap: 5px;
-}
-
-.metaIcon {
-  width: 15px;
-  height: 15px;
-  font-size: 15px;
-  fill: currentColor;
-  opacity: 0.55;
-}
