--- src/components/inference/templates-page.module.css
@@ -1,150 +1,0 @@
-/* Bundles catalogue. Mirrors templates-page.module.css so the two catalogues read as one system;
-   what differs is the per-service grouping (no category/hardware axes — a bundle inherits both from
-   the service it runs on, and the sections already separate them). */
-
-.headerRow {
-  display: flex;
-  flex-wrap: wrap;
-  align-items: flex-end;
-  justify-content: space-between;
-  gap: 16px;
-}
-
-.headerText {
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-  min-width: 260px;
-}
-
-.search {
-  width: 240px;
-}
-
-.searchIcon {
-  margin-right: 4px;
-  color: var(--text-secondary);
-  font-size: 20px;
-}
-
-.searchClearIcon {
-  font-size: 18px;
-}
-
-.summary {
-  color: var(--text-secondary);
-  font-size: 12.5px;
-}
-
-/* One section per parent service. */
-
-.group {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-}
-
-.groupHead {
-  display: flex;
-  align-items: baseline;
-  gap: 10px;
-  padding-bottom: 6px;
-  border-bottom: 1px solid var(--border-glass);
-}
-
-.groupTitle {
-  margin: 0;
-  overflow: hidden;
-  font-size: 15px;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-}
-
-.groupCount {
-  flex: 1;
-  color: var(--text-secondary);
-  font-size: 12px;
-}
-
-.groupLink {
-  flex: 0 0 auto;
-}
-
-.grid {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(430px, 1fr));
-  gap: 8px;
-}
-
-@media (max-width: 560px) {
-  .grid {
-    grid-template-columns: 1fr;
-  }
-}
-
-/* States */
-
-.emptyState {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 10px;
-  padding: 44px 24px;
-  border: 1.5px dashed var(--border);
-  border-radius: 20px;
-  background: var(--background-glass);
-  text-align: center;
-}
-
-.emptyIcon {
-  width: 30px;
-  height: 30px;
-  color: var(--accent1);
-  font-size: 30px;
-  fill: currentColor;
-}
-
-.emptyTitle {
-  font-family: var(--font-plus-jakarta-sans), var(--font-inter), sans-serif;
-  font-size: 16px;
-  font-weight: 700;
-}
-
-.emptyHint {
-  max-width: 420px;
-  color: var(--text-secondary);
-  font-size: 12.5px;
-}
-
-.emptyAction {
-  margin-top: 6px;
-}
-
-.stateBox {
-  padding: 24px 0;
-  text-align: center;
-}
-
-/* Loading skeletons — same grid rhythm as the real cards, so nothing jumps on arrival. */
-
-.skeletonCard {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-  padding: 18px 20px;
-  border: 1px solid var(--border-glass);
-  border-radius: 20px;
-  background: var(--background-glass);
-}
-
-.skeletonTop {
-  display: flex;
-  gap: 12px;
-}
-
-.skeletonLines {
-  display: flex;
-  flex: 1;
-  flex-direction: column;
-  gap: 6px;
-}
